Add dummysource / xgpu averaging & downselect
Dummysource replaces the ethernet input for throughput
testing, and is enabled with the commandline switch --fakesource

Add xGPU averaging and subselection. The former has been "tested" in
that it outputs appropriate data when the pipeline is fed with the
all ones.

With all threads active, the pipeline runs at ~40Gb/s on my
old Xeon machine, seemingly processing limited by my RTX 2060 GPU.

NB: Probably some syncronization barriers are needed, certainly
on the block which copies data to the GPU. See
https://github.com/ledatelescope/bifrost/issues/138