Add triggered dump path
- Add a second copy process to copy data from small to large buffer
- Add placeholder triggered dump process
- Make copy block work with GPU/CPU output rings

It should be possible to "just" make the nethread output buffer parameterizably
large, but it's not obvious [to me] what consequences that would have for the
underlying C packet receiver, which is currently hardcoded to have a 4-gulp buffer.
It's also not obvious what the best way to pass a buffer size down into the receiver would be.