Test with packets
Having verified with the built-in test files, now we move on to
re-benchmarking with packet streams.

It appears that the pipeline doesn't cope well with large header
entries (and json en/decoding of them). Delays in dealing with
these header entries would cause upstream packet loss.
Some of this effect must be user error -- with sufficiently
large copy buffers, a large initial stream creation latency can be
tolerated -- however, it is observed that a ~0.3s call to the bf_xgpu
library in the Correlator block to generate the output baseline order
causes packet loss no matter the depth of the upstream buffer.
Replacing this library call with a 1s sleep (in python) does not cause
this error. It is unclear if the underlying code is not being appropriately
core locked, or there is some other issue.

In any case, this commit avoids these errors by setting input ordering
for the correlator / corr output block at instantiation time. It means
that the pipeline must be completely restarted if the configuration changes.
In practice, this is probably not a particularly problematic requirement.

This commit has been tested at an input data rate of ~27 Gb/s and can
successfully run on rtr-dev2 with no packet loss with the full correlator
pipeline (but not the beamforming or sub-correlator processors) enabled.

Critically, all processing threads -- *AND* their hyperthread siblings
should be kept away from the core handling packet reception.