Start with no UDTs
Start with No UDTs, so that the first attempt to set output
IP addresses results in their creation.
This is much less likely to cause packet drops vs destroying
and recreating the UDTs, which appears to cause upstream issues.
Some empirically determined (but imperfect) sleeps on UDT
destruction/creation have been added to try and solve this issue